Evaluation (#2)

+++ src/chess/types/square.h
@@ -29,7 +29,7 @@
 #include "file.h"
 #include "rank.h"
 
-enum Square : std::int32_t {
+enum Square : std::int8_t {
     A8, B8, C8, D8, E8, F8, G8, H8,
     A7, B7, C7, D7, E7, F7, G7, H7,
     A6, B6, C6, D6, E6, F6, G6, H6,
